force system python instead of homebrew
So it turns out parts of the KA githooks currently break if there is a
python link in /usr/local/bin, due to that path being prepended to the
$PATH during execution, leading to the active virtualenv for the shell
being ignored. So for now we have to ensure homebrew python is NOT
installed, until we figure out the issue with recent versions of git.

This change unfortunately now necessitates the use of `sudo` because of
system python.